+++ unnamed_client/src/main.rs
@@ -3,4 +3,5 @@ use unnamed_engine::core::engine::Engine;
 fn main() {
     let mut engine = Engine::default();
     engine.run();
+    engine.shutdown();
 }